feat: improve RAT client implementation
- Enhance state leaf evidence generation
- Improve state trie synchronization
- Update adjacent leaves submitter
- Add divergence verification tests

+++ clients/rat-client-type3/pkg/verification/opnode_verifier.go
@@ -140,7 +140,7 @@ func (v *OpNodeVerifier) GetSafeL2HeadAtL1(
 		return 0, err
 	}
 
-	return resp.SafeHead.Number, nil
+	return uint64(resp.SafeHead.Number), nil
 }
 
 // GetSyncStatus gets the current sync status from op-node
